Move player_action into player_control module
