+++ xlive/Blam/Engine/Simulation/SimulationWorld.h
@@ -1,7 +1,4 @@
 #pragma once
-#include "Blam/Common/Common.h"
-#include "Blam/Engine/game/aim_assist.h"
-#include "Blam/Math/real_math.h"
 
 enum e_simulation_world_type
 {
@@ -25,27 +22,3 @@ enum e_simulation_world_state
 	_simulation_world_state_leaving = 0x6,
 	k_simulation_world_state_count = 0x7,
 };
-
-#pragma pack(push, 1)
-struct player_action
-{
-	int control_flag0;
-	int control_flag1;
-	real_euler_angles2d facing;
-	real_point2d throttle;
-	float trigger;
-	float secondary_trigger;
-	DWORD action_flags;
-	WORD weapon_set_identifier;
-	BYTE primary_weapon_index;
-	BYTE secondary_weapon_index;
-	WORD grenade_index;
-	WORD zoom_level;
-	int interaction_type;
-	int interaction_object;
-	int melee_target_unit;
-	s_aim_assist_targetting_data aim_assist_data;
-	int unk;
-};
-CHECK_STRUCT_SIZE(player_action, 0x60);
-#pragma pack(pop)
