+++ xlive/Blam/Engine/objects/objects.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "Blam/Math/BlamMath.h"
+#include "Blam/Engine/game/aim_assist.h"
 #include "Blam/Engine/memory/data.h"
 #include "Blam/Engine/Simulation/SimulationWorld.h"
 #include "Blam/Engine/objects/object_placement.h"
